requirements.txt
python-telegram-bot
pydantic
pydantic-settings
asyncpg
beautifulsoup4
httpx
groq

-r requirements-dev.txt
